Safer secret key generation, though less diverse.
This may or may not suit most needs, as it does not include all characters,
old:
  abcdefghijklmnopqrstuvwxyz0123456789!@#$%^*(-_=+)
  (50 characters available, WYSIWYG)
new:
  a-z,A-Z,0-9,+,/
  (64 characters available, RFC 4648 base64 encoding)

So, the base64 version actually has possibility of greater entropy, even though a brute-force attack would have smaller character sets to deal with.  6 of one, half a dozen of the other.

The version introduced here is at least string safe, and cleaner/easier to work with in other shell/config processing/auditing utilitites.

--
Now the use of machine randomness, that's something feel personally guilty comitting.

Signed-off-by: Isaac (.ike) Levy <ike@blackskyresearch.net>
Signed-off-by: Dana Spiegel <dana@sociabledesign.com>